Formal verification: Accept identical netlists as equivalent.
Netlists (`*_gate.v`) are postprocessed to get rid of
non important details: attributes with source file/line references,
and `keep` attributes.

They are then checked for textual equality. If the contents are the
same, equivalence is reported without running yosys. This allows to
avoid some problems where Yosys reports inequivalence just because it
is unable to prove something.