Process log files properly when using PARSER=surelog
The string that is used in sed to cut the interesting part of a log
is not supposed to be the same with all the PARSER options.
With PARSER=surelog it should be "Executing UHDM frontend",
while with PARSER=yosys-plugin it should be "Executing Verilog with UHDM frontend".

This caused the logs to be empty in one configuration.

Signed-off-by: Wojciech Sipak <wsipak@antmicro.com>