Disable lint rules as they are currently creating more noise than signal.
Leave the github action as-is, just disable for now.
test-cases are typically coming from various external sources or are
intentionally not adhering to lint rules as they expose a particular problem.

It is probably more worthwhile to apply lint rules offline than
commenting on pull requests which mostly results in non-actionable
suggestions, increasing noise. Until the details are figured out:
disable this for now.